# backend/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "item": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "date": {
                        "type": "string"
                    },
                    "subcategoryId": {
                        "type": "string"
                    }
                },
                "required": ["id", "item", "amount", "date", "subcategoryId"],
                "type": "object",
                "additionalProperties": false
            },
            "TransactionRequest": {
                "properties": {
                    "item": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "date": {
                        "type": "string"
                    },
                    "subcategoryId": {
                        "type": "string"
                    }
                },
                "required": ["item", "amount", "date", "subcategoryId"],
                "type": "object",
                "additionalProperties": false
            },
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
            "Template": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "date": {
                        "type": "string",
                        "nullable": true
                    },
                    "subcategoryId": {
                        "type": "string"
                    }
                },
                "required": ["id", "name", "itemName", "amount", "date", "subcategoryId"],
                "type": "object",
                "additionalProperties": false
            },
            "TemplateRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "date": {
                        "type": "string",
                        "nullable": true
                    },
                    "subcategoryId": {
                        "type": "string"
                    }
                },
                "required": ["name", "itemName", "amount", "date", "subcategoryId"],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "balanzo-backend",
        "version": "0.0.1",
        "description": "Node.js backend for Balanzo.",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Roland Csősz"
        }
    },
    "paths": {
...
        "/transactions": {
            "post": {
                "operationId": "CreateTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Transactions"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequest"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetTransactions",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            },
                                            "type": "array"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Transactions"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/transactions/{id}": {
            "get": {
                "operationId": "GetTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Transactions"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Transactions"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteTransaction",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Transactions"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/templates": {
            "post": {
                "operationId": "CreateTemplate",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Template"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Templates"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateRequest"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetTemplates",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/Template"
                                            },
                                            "type": "array"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Templates"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/templates/{id}": {
            "get": {
                "operationId": "GetTemplate",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Template"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Templates"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateTemplate",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Templates"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteTemplate",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": ["Templates"],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}